chore(assembler): Added CInstruction struct + bitcode mapping
The current implementation for the C Instruction in the Hack
architecture is quite simple , we just have the 3 different section
each mapped directly to their bits using Go map(s). At this point
the codegen process is a simple get operation on the map plus some
bitwise logic to pack each bitcode in its rightful place.