feat(assembler): Added code generation for C Instructions
The following commits add the CodeGeneraor.TranslateCInst() method to
complete the codegen phase of the Assembler pipeline. The program
defined in the 'main.go' file has been tested with the CPUEmulator and
works as expected, more test will be written later on.
Also, I added an new empty association in the 'DestTable' and 'JumpTable'
since that fields are optional and don't need to trigger an error during
the mapping resolution phase.