feat(assembler): Added AST generation, simplified 'parser.go'
This commit iterates on the work done with 'goparsec' library and
provides a new AST functionaliy instead of the previous raw token lexing
one. This means that before the hack codegen it will be only needed to
traverse the AST and convert each parsed instruction to its own
in-memory version. For label declaration we'll only need to assign them
a value in the symbole table when reached such nodes.