feat(asm): Added draft impl for 'codegen.go' pass
This new codegen pass for the assembler (Asm) taget is the cornerstone
of the VM Translator project, it will take the in memory representation
and spits out a textual one that can be dumped or elaborated further.

I still have some questions about the how strict the checks should be
during the codegeneration phase, nevertheless for the time being I tried
to be as strict as possible, maybe in future I'll remove the TODOs.