feat(vm): Added PCs for VMs MemoryOp and ArithmeticOp statement
The following commit setups the parsing module for the parsing of the VM
intermediate language (project 7 and 8). Mainly here we implemented the
first two basic instruction to load operands from memory and perform
operations on the stack.

I'm still trying to figure out a better solution than extracting the
data about the AST by visiting the tree of pc.Queryable which requires
to check quite pedantically every node before extracting the info.