chore(vm): 'lowering.go' now takes advantage of intermediate register
Now we take advantage of R13 to save the value before pushing the stack,
by the documentation provided on the manual this register are free to be
used by the VM implementation to better manage the .asm translation.

Also to better manage difference between Push and Pop operations I splitted
into two specialized method with the dispatch on 'HandleMemoryOp'.