fix(vm): Fixed a problem in the lowering of 'vm.ReturnOp'
There was a bug that caused the called routine to go to an undefined
location in memory when trying to give back control to the caller by
jumping to its declared return_address. Now it's fixed and while the
test case still fails we see that the segment are setup and restored
correctly when calling (or returning from) a function.