fix(vm): Fixed local variable allocation in 'vm.FuncDecl'
While previously I wiped clean the local segment for the function as
expected I forgot to increment the Stack Pointer, this meant that at
runtime the push and pop operation made a double/triple write on the
local segment causing a sort of double write bug that then caused the
problem in the arithmetic result of the addition(s).

Now the 'NestedCall' suite seems to be working as expected, just one
note: somewhere in the file provided there's written that it possible to
attach freely the bootstrap code to this test and it will work correctly
either way. This is not true, if you inject the bootstrap code the test
will fail simply because the base of the SP is fixed to 261 and after
the call stack unwinding the test expects it to be 261 while the
bootstrap code injected has already rewritten it to 256 (which is
technically correct) but the test will fail and leave you with a
really serious headache, I'll try to send a mail to the team tomorrow