chore(projects): Completed source code study for Project 09
Also, I moved back out of its own folder the BitMap Editor that (for
waht I gathered) is to be used to create sprites and improve graphic
performance of the program. As the program choice I'll probably go for
the classic Tetris game (since the project is called nand2tetris).