feat(jack): Extended parsing support to constructors and methods
This will allow (through the `pSubroitune` parser) to supoprt at the
same type all the three types of subroutine defined in Jack. During
the 2nd phase of the parsing process we will need to do some checks
on each (example: all constructors should return as a type the class
they constructs and its name/identifier should be 'new').