feat(jack): Added support for nested expr and unary expr
The main highlight of this commit is the support to nested parsing
of subexpression (in both binary and unary form), this allows my
parser to support the following syntax: `(a+b) * (b+c)` w/o dying.

To do so without incurring in an infinite recursion problem I had
to split the `pTerm` from the `pExpr`, the idea is to give priority
to "simple to test PCs" like `pIdent` and `pLiteral` and (only if
necessary) yield back control to `pExpr` with a full-on recursive
search. If we went the `pExpr` route always the current implementation
would've run until the stack size was exhausted. This is beccause in
`pExpr` the complex expression  with potential nesting and a recursive
first structure (e.g.  UnaryExpr, BinaryExpr, ...) are evaluated first
while in `pTerm` they are purposefully checked last. This simple trick
stops recursion to only when strictly necessary allowing us to complete
the parsing process w/o consuming excessive amounts of stack.