feat(jack): Added stdlib/OS ABI to not block compilation until Project 12
I've generated a JSON file (stdlib.json) by parsing the AST of the standard
library (taken from project 12 stub files),, since most of the methods are
not implementend we use it only for the function signature.

The following structure will allow every program to be compiled even w/o
the standard library being written, it can be 'linked' only when running
the code on the VM Emulator is needed. Once available we will ignore the
'built-in' ABI if the source code for the stdlib is provided via CLI.