feat(jack): Added ScopeTable to track scopes and variable definitions
This will help centralize the logic for managing scopes, for now just
class and subroutine scopes are supported but it can be extended in future.
Also in each scope the variables declared inside are tracked and registered
in the correct VM memory segment. We may need to add in future checks so that,
for example, Static and Field variables can be declared only in class scope and
not in subroutine scope. The same applies for Local and Argument variables as well.

We incredibly support variables shadowing and correctly clean up (or deallocate) scopes.