chore(jack): Moved from Go's maps to OrderedMaps
TO simplify my life when inspecting diffs of compiled code I need to place
each subroutine exactly in the same order/place they are defined in the Jack
class. This is not possible with Go's maps, given that they are unsorted by
default so I decided to copy 'typescript-go' implementation (minus some features).