fix(jack): Argument declaration is now meaningful for address calculation
Before we just used a map[string]Variable data structure that does not
keep track of insertion order (unsorted hashmap), this meant that in
specific build and/or run of the code the offset of the variable could
change just because they were registered in a different order than the
declared one.

This was an issue because the argument were put on the stack in the
declared order but their offset was not always matching, meaning that at
runtime different addresses were picked up instead of the expected ones.

The fix was trivial: we just added an OrderMap on the 'Subroutine.Arguments'
field instead of the built-in map provided by Go.