test(jack_compiler): Added full test suite for Project 10 and Project 11
Given that there's no test suite provided by the project authors for now
I'm just checking that the compilation succeeds for all the projects/files.
Will try to work on a more definitive solution later that accounts also
for basic regression testing and fails if the produced VM code changes.