test(jack_compiler): Added reproducible builds and no-regression tests
Now that I am able to produce exactly the same output for exactly the
same input I just added a simple no-regression test on all the provided
Jack code of Project 9. That .vm files have been compiled with the
built-in Jack Compiler and not with my own version of it so we can assume
them to be correct.

There are a couple of differences between my codegen and the original
one, mainly in how we handle label randomization and if-then statements.
Each difference has been reviewed and it's cleared personally by me.

The test is quite easy: I created a diff for each project of the changes
from the built-in version to mine, I run my compiler on the source code
and expect after the execution to find exactly the same differences as
the one I created, if not it means there was a regression/change in my code.