chore(projects): Moved around and setup '13 - Jack III: Type Checking'
Since the 13th project is left up to the student, I have decided to
try an implement a typechecking system (and a new compilation pass) for
the Jack programming language. This will be an optional check that can be
enabled via CLI when running the Jack compiler.

As test cases I forked the source code from Project 09 since it seems to
be a wide enough codebase for testing my new typechecker pass.