fix(jack): Fixed all typechecking issues for Project 13
I also removed the diff files for the projects since we don't need nor use
them on this project and specifically at this stage. The output files are
exactly equal to the previous ones given that typechecking doesn't impact
codegen at all. Most of the problem encountered were trivial issues.