chore(jack): Resolved some minor syntax issues on my custom compiler
The compiler provided by the nand2tetris' team is more powerful than the
spec I was given so it could compile a more relaxed version of the Jack
grammar than mine. This can technically be reproduced on mine but is quite
boring so I've just decided to amend the code instead.

The 2 main issue are in nested expression parsing (my compiler always need
parentheses to parse nested expressions) and in comment parsing (in my
compiler they are not allowed/expected anywhere).