feat(projects): Went back to 'Project 9' and drafted first code
I initially skipped this project since it was less interesting to me than
the others where we actually built the compiler and toolchain. After coding
in the constrained environment of the OS/stdlib I decided to get back and
challenge myself in writing a small graphical program with less syntactical
constructs, primitives and abstraction than the one that are usually expected.