feat(projects): Now I'm able to draw an auto-resizable cursor (sort of)
There are some problems where the cursor is not perfectly aligned to the
center of the cell/screen. I think they are related to having to work with
integers (integer division and so on) but I'm not sure. Even so I don't really
give a fuck, I just want to finish this project and move on to the next one.